chore: move files to more appropriate places
