+++ packages/features/data-table/index.ts
@@ -2,7 +2,7 @@ export * from "./components";
 export * from "./lib/types";
 export * from "./lib/utils";
 export * from "./lib/serializers";
+export * from "./lib/separator";
 export * from "./DataTableProvider";
 export * from "./hooks/useColumnResizing";
 export * from "./hooks";
-export * from "./types";
